+++ PFF/src/util/util.cpp
@@ -196,28 +196,28 @@ namespace PFF::util {
 
         SYSTEMTIME win_time;
         GetLocalTime(&win_time);
-        loc_system_time.year = static_cast<unsigned short>(win_time.wYear);
-        loc_system_time.month = static_cast<unsigned short>(win_time.wMonth);
-        loc_system_time.day = static_cast<unsigned short>(win_time.wDay);
-        loc_system_time.day_of_week = static_cast<unsigned short>(win_time.wDayOfWeek);
-        loc_system_time.hour = static_cast<unsigned short>(win_time.wHour);
-        loc_system_time.minute = static_cast<unsigned short>(win_time.wMinute);
-        loc_system_time.secund = static_cast<unsigned short>(win_time.wSecond);
-        loc_system_time.millisecends = static_cast<unsigned short>(win_time.wMilliseconds);
+        loc_system_time.year =              static_cast<u16>(win_time.wYear);
+        loc_system_time.month =             static_cast<u8>(win_time.wMonth);
+        loc_system_time.day =               static_cast<u8>(win_time.wDay);
+        loc_system_time.day_of_week =       static_cast<u8>(win_time.wDayOfWeek);
+        loc_system_time.hour =              static_cast<u8>(win_time.wHour);
+        loc_system_time.minute =            static_cast<u8>(win_time.wMinute);
+        loc_system_time.secund =            static_cast<u8>(win_time.wSecond);
+        loc_system_time.millisecends =      static_cast<u16>(win_time.wMilliseconds);
 
 #elif defined PFF_PLATFORM_LINUX || defined PFF_PLATFORM_MAC
 
         struct timeval tv;
         gettimeofday(&tv, NULL);
         struct tm* ptm = localtime(&tv.tv_sec);
-        loc_system_time.year = static_cast<unsigned short>(ptm->tm_year + 1900);
-        loc_system_time.month = static_cast<unsigned short>(ptm->tm_mon + 1);
-        loc_system_time.day = static_cast<unsigned short>(ptm->tm_mday);
-        loc_system_time.day_of_week = static_cast<unsigned short>(ptm->tm_wday);
-        loc_system_time.hour = static_cast<unsigned short>(ptm->tm_hour);
-        loc_system_time.minute = static_cast<unsigned short>(ptm->tm_min);
-        loc_system_time.secund = static_cast<unsigned short>(ptm->tm_sec);
-        loc_system_time.millisecends = static_cast<unsigned short>(tv.tv_usec / 1000);
+        loc_system_time.year =              static_cast<u16>(ptm->tm_year + 1900);
+        loc_system_time.month =             static_cast<u8>(ptm->tm_mon + 1);
+        loc_system_time.day =               static_cast<u8>(ptm->tm_mday);
+        loc_system_time.day_of_week =       static_cast<u8>(ptm->tm_wday);
+        loc_system_time.hour =              static_cast<u8>(ptm->tm_hour);
+        loc_system_time.minute =            static_cast<u8>(ptm->tm_min);
+        loc_system_time.secund =            static_cast<u8>(ptm->tm_sec);
+        loc_system_time.millisecends =      static_cast<u16>(tv.tv_usec / 1000);
 
 #endif
         return loc_system_time;
